split datasets.go
